requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==6.0.1
djangorestframework==3.16.1
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
sqlparse==0.5.5
